bugfix: avoid truncating Lua context in filters.
The request context was widened from uint16_t to uint32_t when phase
flags grew beyond 16 bits, but the header and body filters continued
saving it in a uint16_t old_context variable.

Contexts above 0xffff were truncated while running a filter and
restored as zero afterward. Use uint32_t for old_context to preserve
the complete phase value.

+++ src/ngx_http_lua_bodyfilterby.c
@@ -241,7 +241,7 @@ ngx_http_lua_body_filter(ngx_http_request_t *r, ngx_chain_t *in)
     ngx_http_lua_loc_conf_t     *llcf;
     ngx_http_lua_ctx_t          *ctx;
     ngx_int_t                    rc;
-    uint16_t                     old_context;
+    uint32_t                     old_context;
     ngx_pool_cleanup_t          *cln;
     ngx_chain_t                 *out;
     ngx_chain_t                 *cl, *ln;
